* Added a new C-only version of the vasan runtime. This one works with musl libc (unlike the previous one, which was crashing musl in one of its constructors).
* The new runtime is once again linked in statically, but since
the symbols are now extern and visible, all calls get routed through
the PLT. This means that we'll end up using the vasan runtime of whichever
lib/binary gets loaded first.

* I've successfully tested this in a simple program with multiple libs
implementing vararg functions. vararg violations get detected, even if
the caller and callee are in different binaries.
Unfortunately, this does not work for printf calls yet. musl's printf
does not get instrumented :( Please find out what is causing this Priyam!

* I've also changed how the VASAN_ERR_LOG_PATH environment variable is
handled. If the env variable is set, the sanitizer goes into logging-only
mode. It logs all violations into the error log, but does not terminate
the program when a violation is detected. If the env var is not set,
errors are logged to stderr and the program is terminated immediately.

+++ llvm/projects/compiler-rt/lib/vasan-old/vasan.h
@@ -1,17 +1,7 @@
-#include <iostream> 
-#include <cxxabi.h>
-#include <string>
-#include <string.h>
-#include <stdio.h>
-#include <csignal>
-#include <signal.h>
-#include <ucontext.h>
-#include <list>
-#include <stack>
-
+#include <cstdint>
 struct callerside_info {
-	uint32_t id;
-	uint32_t arg_count;
+	uint64_t id;
+	uint64_t arg_count;
 	uint64_t*  arg_array;
 	
 };
